Add project setup script
I've added a basic script that just you may run on setup of a new
machine. It just checks for all the things that we went through together
- AWS CLI, CDK, Expo, etc. Nothing wild.

In light of that I have created a new folder for housing scripts we
might write, and moved `fetch_graphql.sh` there.