Add user state to global auth context
This is a temporary fix to the previous issue of not being able to
access the current user from any component within the auth context. It
should be changed such that the `App.tsx` component passes it into the
Auth context, rather than the entire application being wrapped in this.